dev: homwork 12.3
